+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## v0.28.0
+- Added rich CLI w/ command history (#347)
+- Fixed sending Ctrl+C to Windows executable running via WSL
 - Explicitly disabled Content-Encoding for byte range responses
     - If the Range header is present in a request, Content-Encoding is not honored